+++ src/components/VideoPage/VideoPanel/VideoPanel.jsx
@@ -1,11 +1,15 @@
+import { addVideoToLikedVideo } from "../../../services/like-services/addVideoToLikedVideo";
 import "../css/VideoPage.css";
 
 const VideoPanel = () => {
+  const handleLikedVideos = () => {};
   return (
     <div className="video-panel">
       <div className="video-sub-menus">
         <button className="btn btn-icon">
-          <span className="material-icons-outlined">thumb_up</span>
+          <span className="material-icons-outlined" onClick={handleLikedVideos}>
+            thumb_up
+          </span>
           Like
         </button>
         <button className="btn btn-icon">
