+++ src/pages/VideoListingPage/VideoListingPage.jsx
@@ -1,28 +1,15 @@
 import "../VideoListingPage/VideoListingPage.css";
 import { CategoryChips, VideoCard } from "../../components";
-import axios from "axios";
-import { useEffect, useState } from "react";
+
+import { useEffect } from "react";
 import { useData } from "../../context/dataContext/dataContext";
 
 const VideoListingPage = () => {
-  const [video, setVideo] = useState([]);
-  const { clickedCategory } = useData();
+  const { clickedCategory, getAllVideosFromDatabase, video } = useData();
 
   useEffect(() => {
-    (async () => {
-      const response = await axios.get("/api/videos");
-      const data = response.data.videos;
-      if (clickedCategory !== "All") {
-        const filterData = [...data].filter(
-          (item) => item.category === clickedCategory
-        );
-        setVideo(filterData);
-        return filterData;
-      }
-      setVideo(data);
-      return data;
-    })();
-  }, [clickedCategory]);
+    getAllVideosFromDatabase();
+  }, [clickedCategory, getAllVideosFromDatabase]);
 
   return (
     <div className="main-content-page video-listing-page">
