+++ src/services/like-services/addVideoToLikedVideo.jsx
@@ -1,0 +1,13 @@
+import axios from "axios";
+
+const addVideoToLikedVideo = (authToken, video) => {
+  axios.post(
+    "/api/user/likes",
+    { video },
+    {
+      headers: { authorization: authToken }
+    }
+  );
+};
+
+export { addVideoToLikedVideo };
